pyproject.toml
[tool.poetry]
name = "new-rest-api"
version = "0.1.0"
description = ""
authors = ["andrechizh8 <andrechizh.ru@yandex.ru>"]
readme = "README.md"
packages = [{include = "new_rest_api"}]

[tool.poetry.dependencies]
python = "^3.9"
selene = "2.0.0b16"
pytest = "^7.2.1"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
requests = "^2.28.2"
allure-pytest = "^2.12.0"
python-dotenv = "^0.21.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
